requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0

langgraph>=0.0.10

# LLM Providers
openai>=1.3.0
ollama>=0.1.0
anthropic>=0.18.0

# packaging
setuptools
build